Updated Questions_Answers to get data from API when component mounts
