+++ Client/Questions_Answers/Questions_Answers.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import axios from 'axios';
 import Questions from './Questions.js';
 import Search from './Search.js';
 
@@ -7,11 +8,65 @@ class Questions_Answers extends React.Component {
     super(props);
 
     this.state = {
-      questions: props.questions,
-      displayedQuestions: props.questions.results.slice(0, 2),
+      questions: [],
+      displayedQuestions: [],
       searchText: ''
     };
 
+  }
+
+  componentDidMount () {
+    var orderAnswers = (ans) => {
+      let orderedAnswers = [];
+
+      for (var key in ans) {
+        if (orderedAnswers.length === 0) {
+          orderedAnswers.push(ans[key]);
+
+        } else {
+          if (ans[key].helpfulness >= orderedAnswers[0].helpfulness) {
+            orderedAnswers.unshift(ans[key]);
+
+          } else if (ans[key].helpfulness <= orderedAnswers[0].helpfulness) {
+            orderedAnswers.push(ans[key]);
+          }
+        }
+      }
+
+      // if answerer_name === Seller, it needs to be put on the top of the array
+      // maybe that could be done in the top iteration? put sellers into a separate array and then merge it once others have been sorted
+
+      orderedAnswers.sort((a, b) => {
+        return b.helpfulness - a.helpfulness;
+      })
+
+      return orderedAnswers;
+    }
+
+    var randomIndex = Math.floor(Math.random() * 1011);
+    randomIndex += 64620;
+
+    axios.post('/questions', {
+      productId: randomIndex
+    })
+    .then((response) => {
+      console.log('Successful Question Request: ', response.data);
+
+      response.data.results.forEach((question) => {
+        question.answers = orderAnswers(question.answers)
+      })
+
+      return response.data;
+    })
+    .then((questions) => {
+      this.setState({
+        questions: questions,
+        displayedQuestions: questions.results.slice(0, 2)
+      })
+    })
+    .catch((error) => {
+      console.log(`There was an error getting question data: ${error}`);
+    })
   }
 
   onSearchChange (e) {
